uint32_t

+++ src/os400qc3.c
@@ -460,7 +460,7 @@ int ssh2_bn_from_bin(ssh2_bn *bn, size_t len, const unsigned char *val)
     return 0;
 }
 
-int ssh2_bn_set_word(ssh2_bn *bn, unsigned long val)
+int ssh2_bn_set_word(ssh2_bn *bn, uint32_t val)
 {
     val = htonl(val);
     return ssh2_bn_from_bin(bn, sizeof(val), (unsigned char *)&val);
